+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - RESTFul and MQTT config refactor
 - stability and performance improvements
 - added generator support
+- added generator state sensor
+- added generator mode sensor
+- added generator mode control
 - added data refresh control direct from home app as a extra tile, native app
 - added data refresh control direct from envoy section, 3rd party app
 - added data refresh sensor